% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{say_hello}
\alias{say_hello}
\title{Hello function}
\usage{
say_hello(name)
}
\arguments{
\item{name}{Character value of persons name to say hello to}
}
\value{
Returns a character vector saying hello
}
\description{
A short description...
}
\examples{
say_hello("Colin")
say_hello("Jane Doe")


}
